Move avatar utility funcitons into avatar-utils
