+++ src/assets/avatars/avatars.js
@@ -1,5 +1,3 @@
-import { getReticulumFetchUrl } from "../../utils/phoenix-utils";
-
 export const avatars = [
   {
     id: "botdefault",
@@ -42,28 +40,3 @@ export const avatars = [
     model: "https://asset-bundles-prod.reticulum.io/bots/BotWoody_Avatar-0140485a23.gltf"
   }
 ];
-
-export const AVATAR_TYPES = {
-  LEGACY: "legacy",
-  SKINNABLE: "skinnable",
-  URL: "url"
-};
-
-const legacyAvatarIds = avatars.map(a => a.id);
-export function getAvatarType(avatarId) {
-  if (!avatarId || legacyAvatarIds.indexOf(avatarId) !== -1) return AVATAR_TYPES.LEGACY;
-  if (avatarId.startsWith("http")) return AVATAR_TYPES.URL;
-  return AVATAR_TYPES.SKINNABLE;
-}
-
-export async function getAvatarSrc(avatarId) {
-  switch (getAvatarType(avatarId)) {
-    case AVATAR_TYPES.LEGACY:
-      return `#${avatarId}`;
-    case AVATAR_TYPES.SKINNABLE:
-      return fetch(getReticulumFetchUrl(`/api/v1/avatars/${avatarId}`))
-        .then(r => r.json())
-        .then(({ avatars }) => avatars[0].gltf_url);
-  }
-  return avatarId;
-}
