+++ src/components/ik-controller.js
@@ -1,6 +1,6 @@
 const { Vector3, Quaternion, Matrix4, Euler } = THREE;
 
-import { AVATAR_TYPES } from "../assets/avatars/avatars";
+import { AVATAR_TYPES } from "../utils/avatar-utils";
 
 function quaternionAlmostEquals(epsilon, u, v) {
   // Note: q and -q represent same rotation
